Fix truncate_middle UTF-8 panic, normalize_uint range, and parser RParen
- Use char-based iteration in truncate_middle instead of byte slicing to
  prevent panics on multi-byte UTF-8 characters
- Widen normalize_uint from u64 to u128 to handle large gas/value fields
- Validate closing RParen in parenthesized RPC args, returning an error
  instead of silently accepting incomplete input
- Remove unused ParseError::ExpectedMethod variant

+++ tooling/repl/src/commands/mod.rs
@@ -203,7 +203,7 @@ fn normalize_uint(s: &str) -> Result<String, String> {
     if s.starts_with("0x") {
         // Already hex
         Ok(s.to_string())
-    } else if let Ok(n) = s.parse::<u64>() {
+    } else if let Ok(n) = s.parse::<u128>() {
         Ok(format!("0x{n:x}"))
     } else {
         Err(format!("invalid uint: {s}"))
